# Conflicts:
#	common/app/pom.xml
#	common/bundle/pom.xml
#	common/pom.xml
#	gsactivities/app/pom.xml
#	gsactivities/bundle/pom.xml
#	gsactivities/pom.xml
#	gsusa/app/pom.xml
#	gsusa/bundle/pom.xml
#	gsusa/pom.xml
#	gsusa/selective-clean/pom.xml
#	pom.xml
#	vtk/app/pom.xml
#	vtk/beanutils-wrapper/pom.xml
#	vtk/bundle/pom.xml
#	vtk/google-collections-wrapper/pom.xml
#	vtk/googleInject-wrapper/pom.xml
#	vtk/ical4j-wrapper/pom.xml
#	vtk/modified-checker-bundle/pom.xml
#	vtk/pom.xml
#	vtk/replication-bundle/pom.xml
#	vtk/scaffolding-bundle/pom.xml
#	web/app/pom.xml
#	web/bundle/pom.xml
#	web/opencsv-wrapper/pom.xml
#	web/pdfbox-wrapper/pom.xml
#	web/pom.xml

+++ vtk/bundle/src/main/java/org/girlscouts/vtk/osgi/service/impl/GirlScoutsTroopOCMServiceImpl.java
@@ -32,16 +32,24 @@ public class GirlScoutsTroopOCMServiceImpl implements GirlScoutsTroopOCMService 
 
     @Override
     public Troop create(Troop object) {
-        TroopNode node = NodeToModelMapper.INSTANCE.toNode(object);
-        node = girlScoutsOCMRepository.create(node);
-        return NodeToModelMapper.INSTANCE.toModel(node);
+        if(!object.getIsTransient()) {
+            TroopNode node = NodeToModelMapper.INSTANCE.toNode(object);
+            node = girlScoutsOCMRepository.create(node);
+            return NodeToModelMapper.INSTANCE.toModel(node);
+        }else{
+            return object;
+        }
     }
 
     @Override
     public Troop update(Troop object) {
-        TroopNode node = NodeToModelMapper.INSTANCE.toNode(object);
-        node = girlScoutsOCMRepository.update(node);
-        return NodeToModelMapper.INSTANCE.toModel(node);
+        if(!object.getIsTransient()) {
+            TroopNode node = NodeToModelMapper.INSTANCE.toNode(object);
+            node = girlScoutsOCMRepository.update(node);
+            return NodeToModelMapper.INSTANCE.toModel(node);
+        }else{
+            return object;
+        }
     }
 
     @Override
@@ -54,7 +62,11 @@ public class GirlScoutsTroopOCMServiceImpl implements GirlScoutsTroopOCMService 
 
     @Override
     public boolean delete(Troop object) {
-        return girlScoutsOCMRepository.delete(NodeToModelMapper.INSTANCE.toNode(object));
+        if(!object.getIsTransient()) {
+            return girlScoutsOCMRepository.delete(NodeToModelMapper.INSTANCE.toNode(object));
+        }else{
+            return false;
+        }
     }
 
     @Override
